Fixed reviewer's marks
