+++ imageFeed/Models/OAuth2TokenResponseBodyModel.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct OAuthTokenResponseBody: Decodable {
+    let accessToken: String
+    let tokenType: String
+    let scope: String
+    let createdAt: Int
+}
