+++ imageFeed/Models/ProfileImageModel.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+struct UserResult: Codable {
+    let profileImage: ProfileImage
+}
+struct ProfileImage: Codable {
+    let small: String?
+    let medium: String?
+    let large: String?
+}
