+++ imageFeed/Profile/ProfileService.swift
@@ -1,39 +1,16 @@
 import Foundation
 import UIKit
 
-struct ProfileResult: Codable {
-    let username: String
-    let firstName: String
-    let lastName: String?
-    let bio: String?
-}
-struct Profile {
-    let username: String
-    let name: String
-    let loginName: String
-    let bio: String?
-    
-    init(profileResult: ProfileResult) {
-        self.username = profileResult.username
-        self.name = "\(profileResult.firstName) \(profileResult.lastName ?? "")"
-        self.loginName = "@\(profileResult.username)"
-        self.bio = profileResult.bio
-    }
-}
-
 final class ProfileService {
     // MARK: - Properties:
     static let shared = ProfileService()
     private let urlSession = URLSession.shared
-    private let builder: URLRequestBuilder
+    private let builder = URLRequestBuilder.shared
     
     private var task: URLSessionTask?
     private var lastToken: String?
     
     private (set) var profile: Profile?
-    init(builder: URLRequestBuilder = .shared) {
-        self.builder = builder
-    }
     // MARK: - Methods:
     func fetchProfile(completion: @escaping (Result<Profile,Error>) -> Void) {
         assert(Thread.isMainThread)
