+++ imageFeed/SpashViewController/SplashViewController.swift
@@ -79,9 +79,9 @@ final class SplashViewController: UIViewController {
         }
     }
     private func showAlert(title: String, message: String) {
-            let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
-            let alertAction = UIAlertAction(title: "OK", style: .default)
-            alert.addAction(alertAction)
+        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
+        let alertAction = UIAlertAction(title: "OK", style: .default)
+        alert.addAction(alertAction)
     }
     
     private func showAuthController() {
